% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleITK2ANTsR.r
\name{antsImage2sitk}
\alias{antsImage2sitk}
\title{convert an antsImage to a SimpleITK image}
\usage{
antsImage2sitk(image)
}
\arguments{
\item{image}{object of class antsImage}
}
\value{
returns a SimpleITK image
}
\description{
convert an antsImage to a SimpleITK image
}

