% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/newAntsExample.r
\name{createAntsArgs}
\alias{createAntsArgs}
\title{interface to generate arguments for antsRegistration command}
\usage{
createAntsArgs(reference, target, setting = "custom", percent = 0.1,
  synpercent = 0.1, searchradius = 4, mattesbins = 32,
  its = "10000x10000x10000", synargs = "[100x10x1,0,5]", cc = FALSE,
  itkthreads = parallel::detectCores(), trans = T, rigid = T,
  similarity = T, affine = T, BSplineDisp = F, gauss = F, expo = F,
  syn = T, mattesweight = 0.5, ccweight = 0.5, dims = 3,
  initTransform = NULL, PSE = NULL)
}
\arguments{
\item{reference}{path to moving image}

\item{target}{path to fixed image}

\item{setting}{name of the setting to be attached to output}

\item{percent}{parameter for ants}

\item{searchradius}{searchradius for CC metric}

\item{mattesbins}{binsize for Mattes MI metric}

\item{its}{parameters for non-syn registration}

\item{synargs}{parameters for syn registration}

\item{itkthreads}{integer: specify number of threads to be used}

\item{trans}{run translation}

\item{rigid}{run rigid matching}

\item{affine}{run affine matching}

\item{BSplineDisp}{use B-Spline deformations}

\item{gauss}{use gaussian Smoothed deformation fields}

\item{expo}{use Exponential deformation type}

\item{syn}{run syn matching}

\item{mattesweight}{weight to attribute to mattes metric}

\item{ccweight}{weight to attribute to CC metric}

\item{dims}{dimensionality of data}

\item{initTransform}{run initial coarse alignment. These features include using the geometric center of the
images (=0), the image intensities (=1), or the origin of the images (=2).}

\item{PSE}{not supported yet}
}
\value{
returns a list with
\item{antsargs}{arguments passed to antsRegistration}
\item{outname}{basename of outputname}
}
\description{
interface to generate arguments for antsRegistration command
}

