% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getAffine.r
\name{getAffineMat}
\alias{getAffineMat}
\title{convert an affine transform file, generated by ANTsR to a 4x4 transformation matrix}
\usage{
getAffineMat(file, IJK2RAS = diag(c(-1, -1, 1, 1)))
}
\arguments{
\item{file}{affine transform file, generated by ANTsR}

\item{IJK2RAS}{IJK to RAS transform}
}
\value{
4x4 matrix
}
\description{
convert an affine transform file, generated by ANTsR to a 4x4 transformation matrix
}

