% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itksnap.r
\name{itksnap}
\alias{itksnap}
\title{call itksnap to view an antsImage}
\usage{
itksnap(image, overlay = NA)
}
\arguments{
\item{image}{antsImage}

\item{overlay}{antsImage as overlay}
}
\description{
call itksnap to view an antsImage
}

