% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleITK2ANTsR.r
\name{sitk2antsImage}
\alias{sitk2antsImage}
\title{convert an  SimpleITK image to an antsImage}
\usage{
sitk2antsImage(image)
}
\arguments{
\item{image}{object of class SimpleITK}
}
\value{
returns a antsImage
}
\description{
convert an  SimpleITK image to an antsImage
}
