% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/newAntsExample.r
\name{antsTransformPoints}
\alias{antsTransformPoints}
\alias{antsTransformPoints.matrix}
\alias{antsTransformPoints.mesh3d}
\title{calls antsApplyTransformsToPoints}
\usage{
antsTransformPoints(mat, affine, warp, antsdir = "~/GIT/DEV/ANTSbuild/bin/",
  ...)

\method{antsTransformPoints}{matrix}(mat, affine, warp,
  antsdir = "~/GIT/DEV/ANTSbuild/bin/", ...)

\method{antsTransformPoints}{mesh3d}(mat, affine, warp,
  antsdir = "~/GIT/DEV/ANTSbuild/bin/", ...)
}
\arguments{
\item{mat}{a matrix or mesh3d}

\item{affine}{path to affine transformation}

\item{warp}{path to warp}

\item{antsdir}{path to directory containing antsApplyTransformsToPoints}
}
\description{
calls antsApplyTransformsToPoints
}

