% Generated by roxygen2 (4.0.1): do not edit by hand
\name{antsTransformPoints}
\alias{antsTransformPoints}
\alias{antsTransformPoints.matrix}
\title{calls antsApplyTransformsToPoints}
\usage{
antsTransformPoints(mat, affine, warp, antsdir = "~/GIT/DEV/ANTSbuild/bin/")

\method{antsTransformPoints}{matrix}(mat, affine, warp,
  antsdir = "~/GIT/DEV/ANTSbuild/bin/")
}
\arguments{
\item{mat}{a matrix or mesh3d}

\item{affine}{path to affine transformation}

\item{warp}{path to warp}
}
\description{
calls antsApplyTransformsToPoints
}

