% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/newAntsExample.r
\name{antsTransformPoints}
\alias{antsTransformPoints}
\alias{antsTransformPoints.matrix}
\alias{antsTransformPoints.mesh3d}
\title{calls antsApplyTransformsToPoints}
\usage{
antsTransformPoints(mat, affine, warp, antsdir = "~/GIT/DEV/ANTSbuild/bin/",
  IJK2RAS = diag(c(-1, -1, 1, 1)), ...)

\method{antsTransformPoints}{matrix}(mat, affine, warp,
  antsdir = "~/GIT/DEV/ANTSbuild/bin/", IJK2RAS = diag(c(-1, -1, 1, 1)),
  ...)

\method{antsTransformPoints}{mesh3d}(mat, affine, warp,
  antsdir = "~/GIT/DEV/ANTSbuild/bin/", IJK2RAS = diag(c(-1, -1, 1, 1)),
  ...)
}
\arguments{
\item{mat}{a matrix or mesh3d}

\item{affine}{path to affine transformation}

\item{warp}{path to warp}

\item{antsdir}{path to directory containing antsApplyTransformsToPoints}

\item{IJK2RAS}{4x4 matrix containing the transform between image and pointspace. For 2D pointclouds this must be the transformation for cbind(mat,0) - the z-coordinate supplemented with zeros.}
}
\description{
calls antsApplyTransformsToPoints
}

