% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsTransformPoints.r
\name{antsTransformPoints}
\alias{antsTransformPoints}
\alias{antsTransformPoints.matrix}
\alias{antsTransformPoints.mesh3d}
\title{a handy wrapper for antsApplyTransformsToPoints}
\usage{
antsTransformPoints(mat, transformlist, IJK2RAS = diag(c(-1, -1, 1, 1)),
  ...)

\method{antsTransformPoints}{matrix}(mat, transformlist,
  whichtoinvert = NA, IJK2RAS = diag(c(-1, -1, 1, 1)), ...)

\method{antsTransformPoints}{mesh3d}(mat, transformlist,
  whichtoinvert = NA, IJK2RAS = diag(c(-1, -1, 1, 1)), ...)
}
\arguments{
\item{mat}{a matrix or mesh3d}

\item{IJK2RAS}{4x4 matrix containing the transform between image and pointspace. For 2D pointclouds this must be the transformation for cbind(mat,0) - the z-coordinate supplemented with zeros.}

\item{affine}{path to affine transformation}

\item{warp}{path to warp}

\item{antsdir}{path to directory containing antsApplyTransformsToPoints}
}
\description{
a handy wrapper for antsApplyTransformsToPoints
}
