% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createAntsArgs}
\alias{createAntsArgs}
\title{interface to generate arguments for antsRegistration command}
\usage{
createAntsArgs(reference, target, setting = "custom", percent = 0.1,
  its = "10000x0x0", synargs = "[100x10x1,0,5]", cc = FALSE,
  itkthreads = 4, trans = T, rigid = T, affine = T, syn = T)
}
\description{
interface to generate arguments for antsRegistration command
}

