% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createAntsArgs}
\alias{createAntsArgs}
\title{interface to generate arguments for antsRegistration command}
\usage{
createAntsArgs(reference, target, setting = "custom", percent = 0.1,
  searchradius = 4, mattesbins = 32, its = "10000x10000x10000",
  synargs = "[100x10x1,0,5]", cc = FALSE, itkthreads = 4, trans = T,
  rigid = T, affine = T, gauss = F, expo = F, syn = T)
}
\arguments{
\item{reference}{path to moving image}

\item{target}{path to fixed image}

\item{setting}{name of the setting to be attached to output}

\item{percent}{parameter for ants}

\item{searchradius}{searchradius for CC metric}

\item{mattesbins}{binsize for Mattes MI metric}

\item{its}{parameters for non-syn registration}

\item{synargs}{parameters for syn registration}

\item{itkthreads}{integer: specify number of threads to be used}

\item{trans}{run translation}

\item{rigid}{run rigid matching}

\item{affine}{run affine matching}

\item{syn}{run syn matching}
}
\description{
interface to generate arguments for antsRegistration command
}

