% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/newAntsExample.r
\name{createAntsArgs}
\alias{createAntsArgs}
\title{interface to generate arguments for antsRegistration command}
\usage{
createAntsArgs(reference, target, setting = "custom", percent = 0.1,
  synpercent = 0.1, searchradius = 4, mattesbins = 32,
  its = "10000x10000x10000", synargs = "[100x10x1,0,5]",
  affine = c("trans", "rigid", "similarity", "affine"),
  elastic = "SyN[0.2,3,0]", elasticmetrics = c("mattes", "cc"),
  metricweights = c(0.5, 0.5), metricreach = c(32, 4), dims = 3,
  initTransform = NULL, elasticS = "3x1x0vox", elasticF = "3x2x1",
  itkthreads = parallel::detectCores())
}
\arguments{
\item{reference}{path to moving image}

\item{target}{path to fixed image}

\item{setting}{name of the setting to be attached to output}

\item{percent}{parameter for ants}

\item{searchradius}{searchradius for CC metric}

\item{mattesbins}{binsize for Mattes MI metric}

\item{its}{parameters for non-syn registration}

\item{synargs}{character: convergence valzes MxNxO [MxNxO,<convergenceThreshold=1e-6>,<convergenceWindowSize=10>]}

\item{affine}{character vector of non-elastic transform to use}

\item{elastic}{character vector of elastic procedures (including parameters). E.g "SyN[0.2,3,0]"}

\item{elasticmetrics}{character vector containing the metrics to be used in elastic matching}

\item{metricweights}{character vector of same length as \code{elasticmetrics}, assigning weights to each metric.}

\item{metricreach}{character vector of same length as \code{elasticmetrics}, assigning the radius/bins for each metric to consider.}

\item{dims}{dimensionality of data}

\item{initTransform}{run initial coarse alignment. These features include using the geometric center of the
images (=0), the image intensities (=1), or the origin of the images (=2).}

\item{elasticS}{smoothing sigmas (see command args of antsRegistration)}

\item{elasticF}{shrink factors (see command args of antsRegistration)}

\item{itkthreads}{integer: specify number of threads to be used}
}
\value{
returns a list with
\item{antsargs}{arguments passed to antsRegistration}
\item{outname}{basename of outputname}
}
\description{
interface to generate arguments for antsRegistration command
}

