% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createAntsArgs.r
\name{createAntsArgs}
\alias{createAntsArgs}
\title{interface to generate arguments for antsRegistration command}
\usage{
createAntsArgs(reference, target, setting = "custom", percent = 0.1,
  affine = c("trans", "rigid", "similarity", "affine"), affinereach = 32,
  affineconverge = "[10000x10000x10000,1e-8,20]", elastic = "SyN[0.2,3,0]",
  elasticpercent = 0.1, elasticconverge = "[100x10x1,0,5]",
  elasticmetrics = c("mattes", "cc"), metricweights = NULL,
  metricreach = c(32, 4), dims = 3, elasticS = "3x1x0vox",
  elasticF = "3x2x1", initTransform = NULL,
  itkthreads = parallel::detectCores(), masks = NULL, folder = NULL,
  transimg = FALSE, verbose = FALSE)
}
\arguments{
\item{reference}{path to moving image}

\item{target}{path to fixed image}

\item{setting}{name of the setting to be attached to output}

\item{percent}{subsampling for affine transforms}

\item{affine}{character vector of non-elastic transform to use}

\item{affinereach}{character vector assigning the bins for Mattes metric used for affine transforms.}

\item{elastic}{character vector of elastic procedures (including parameters). E.g "SyN[0.2,3,0]"}

\item{elasticmetrics}{character vector containing the metrics to be used in elastic matching}

\item{metricweights}{character vector of same length as \code{elasticmetrics}, assigning weights to each metric. If NULL, all selected metrics will be weighted equally.}

\item{metricreach}{character vector of same length as \code{elasticmetrics}, assigning the radius/bins for each metric to consider.}

\item{dims}{dimensionality of data}

\item{elasticS}{smoothing sigmas (see command args of antsRegistration)}

\item{elasticF}{shrink factors (see command args of antsRegistration)}

\item{initTransform}{run initial coarse alignment. These features include using the geometric center of the
images (=0), the image intensities (=1), or the origin of the images (=2).}

\item{itkthreads}{integer: specify number of threads to be used}

\item{folder}{character defining the path where to store the transforms}

\item{transimg}{logical: if TRUE the *diff and *inv will be created.}

\item{verbose}{logcal: if TRUE output are written to console}

\item{affineconvergence}{character: convergence values for affine transforms MxNxO [MxNxO,<convergenceThreshold=1e-6>,<convergenceWindowSize=10>]}

\item{elasticconvergence}{character: convergence valzes MxNxO [MxNxO,<convergenceThreshold=1e-6>,<convergenceWindowSize=10>]}
}
\value{
returns a list with
\item{antsargs}{arguments passed to antsRegistration}
\item{outname}{basename of outputname}
\item{transforms}{list containing characters naming the transforms}
}
\description{
interface to generate arguments for antsRegistration command
}

