% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signedDist.r
\name{DistanceImage}
\alias{DistanceImage}
\title{Create a (signed) distance image from surface mesh}
\usage{
DistanceImage(x, spacing = rep(1, 3), margin = 0.05, sign = TRUE, k = 1,
  IJK2RAS = diag(c(-1, -1, 1, 1)))
}
\arguments{
\item{x}{surface mesh}

\item{spacing}{spacing of grid}

\item{margin}{percentage to add to grid}

\item{sign}{logical: if TRUE signed distances are computed}

\item{k}{k- closest faces to consider during closest point search}

\item{IJK2RAS}{transformation between physical and image space}
}
\value{
returns a distance image of class antsImage
}
\description{
Create a (signed) distance image from surface mesh
}

