% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isosurface.r
\name{isosurface}
\alias{isosurface}
\alias{isosurface._p_itk__simple__Image}
\alias{isosurface.antsImage}
\title{Create isosurface from 3D images}
\usage{
isosurface(x, threshold, as.int = TRUE, ...)

\method{isosurface}{antsImage}(x, threshold, as.int = TRUE, ...)

\method{isosurface}{_p_itk__simple__Image}(x, threshold, as.int = TRUE, ...)
}
\arguments{
\item{x}{image of class antsImage or SimpleITK (_p_itk__simple__Image).}

\item{threshold}{threshold to create isosurface}

\item{as.int}{logical: treat image grey values as integers.}

\item{...}{not used.}
}
\value{
returns a triangulated mesh of class mesh3d.
}
\description{
Create isosurface from 3D images
}

