% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarkTransform.R
\name{landmarkTransform}
\alias{landmarkTransform}
\title{Create an itk transform from landmarks (rigid, similarity and affine)}
\usage{
landmarkTransform(lmFix, lmMoving, type = c("rigid", "similarity",
  "affine"), IJK2RAS = diag(c(-1, -1, 1, 1)[seq_len(ncol(lmFix) + 1)]),
  file = NULL, ...)
}
\arguments{
\item{lmFix}{fix landmarks}

\item{lmMoving}{moving landmarks}

\item{type}{of transform (can be "rigid","similarity" or "affine")}

\item{IJK2RAS}{transform from point to image space.}

\item{file}{filename to save the transform to.}

\item{...}{additional parameters (such as landmark weights) to be passed to \code{\link{rotonto}}, in case type != affine.}
}
\value{
writes a transform to a tempfile and returns the path to that transform
}
\description{
Create an itk transform from landmarks (rigid, similarity and affine) to transform an image
}
\details{
lmMoving are landmarks placed on the image to be transformed to the position of lmFix
}
