% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signedDist.r
\name{Mesh2Image}
\alias{Mesh2Image}
\title{Create a (optionally) binary image from surface mesh}
\usage{
Mesh2Image(x, spacing = rep(1, 3), margin = 0.05, sign = FALSE, k = 1,
  IJK2RAS = diag(c(-1, -1, 1, 1)), inval = 255, outval = 0,
  thresh = min(spacing), binary = FALSE)
}
\arguments{
\item{x}{surface mesh}

\item{spacing}{spacing of grid}

\item{margin}{percentage to add to grid}

\item{sign}{logical: if TRUE signed distances are computed}

\item{k}{k- closest faces to consider during closest point search}

\item{IJK2RAS}{transformation between physical and image space}

\item{inval}{inside value}

\item{outval}{outside value (only considered if binary = TRUE) otherwise outval = thresh.}

\item{thresh}{maximal distance to consider}

\item{binary}{logical: if TRUE the output will be a binary image}
}
\value{
returns a distance image of class antsImage
}
\description{
Create a (optionally) binary image from surface mesh
}

