% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newAntsExample.r
\name{newAntsExample}
\alias{newAntsExample}
\title{This is simply the script newAntsExample.sh as an R function}
\usage{
newAntsExample(reference, target, setting = c("fast", "production"),
  itkthreads = 4)
}
\arguments{
\item{reference}{path to moving image}

\item{target}{path to fixed image}

\item{setting}{presets options ar e fast and production}

\item{itkthreads}{integer: specify number of threads to be used}
}
\description{
This is simply the script newAntsExample.sh as an R function
}
