% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/antsPoints2LabelImage.r
\name{points2LabelImage}
\alias{points2LabelImage}
\title{create a label image from landmark coordinates}
\usage{
points2LabelImage(points, neighbours = 16, spacing = rep(1, 3),
  margin = 0.05, IJK2RAS = diag(c(-1, -1, 1, 1)))
}
\arguments{
\item{neighbours}{integer: number of closest points in the image grid to label}

\item{spacing}{spacing of the output image}

\item{margin}{margin around the landmarks (determined by the extend of the bounding box) to be part of the image}

\item{IJK2RAS}{transform from point to image space.}

\item{pts}{k x 2 or k x 3 matrix containing landmark coordinates}

\item{labelcolors}{vector of length nrow(pts) with a color for each landmark label}
}
\value{
returns an antsImage
}
\description{
create a label image from landmark coordinates
}

